+++ src/main/MQ2CleanUI.cpp
@@ -25,15 +25,26 @@ public:
 	void CleanUI_Trampoline();
 	void CleanUI_Detour()
 	{
-		Benchmark(bmPluginsCleanUI, DebugTry(PluginsCleanUI()));
-		DebugTry(CleanUI_Trampoline());
+		{
+			MQScopedBenchmark bm(bmPluginsCleanUI);
+			PluginsCleanUI();
+		}
+
+		CleanUI_Trampoline();
 	}
 
 	void ReloadUI_Trampoline(bool, bool);
 	void ReloadUI_Detour(bool UseINI, bool bUnknown)
 	{
-		DebugTry(ReloadUI_Trampoline(UseINI, bUnknown));
-		Benchmark(bmPluginsReloadUI, DebugTry(PluginsReloadUI()));
+		ReloadUI_Trampoline(UseINI, bUnknown);
+
+		InitializeInGameUI();
+
+		{
+			MQScopedBenchmark bm(bmPluginsReloadUI);
+
+			PluginsReloadUI();
+		}
 	}
 };
 
@@ -50,14 +61,6 @@ void DrawNetStatus_Detour(uint16_t x, uint16_t y, void* udpConnection, uint32_t 
 
 	DrawNetStatus_Trampoline(x, y, udpConnection, bps);
 	Benchmark(bmPluginsDrawHUD, PluginsDrawHUD());
-
-	// just do this check once. it won't change during execution.
-	static HMODULE hmEQPlayNice = GetModuleHandle("EQPlayNice.dll");
-	if (hmEQPlayNice)
-	{
-		if (fMQPulse pEQPlayNicePulse = (fMQPulse)GetProcAddress(hmEQPlayNice, "Compat_DrawIndicator"))
-			pEQPlayNicePulse();
-	}
 }
 
 void DrawHUD()
@@ -74,13 +77,6 @@ void DrawHUD()
 			}
 
 			Benchmark(bmPluginsDrawHUD, PluginsDrawHUD());
-
-			static HMODULE hmEQPlayNice = GetModuleHandle("EQPlayNice.dll");
-			if (hmEQPlayNice)
-			{
-				if (fMQPulse pEQPlayNicePulse = (fMQPulse)GetProcAddress(hmEQPlayNice, "Compat_DrawIndicator"))
-					pEQPlayNicePulse();
-			}
 		}
 		else
 		{
@@ -170,6 +166,11 @@ void InitializeDisplayHook()
 
 	AddCommand("/netstatusxpos", Cmd_NetStatusXPos);
 	AddCommand("/netstatusypos", Cmd_NetStatusYPos);
+
+	if (GetGameState() == GAMESTATE_INGAME)
+	{
+		InitializeInGameUI();
+	}
 }
 
 void ShutdownDisplayHook()